requirements.txt
pandas
numpy
python-dotenv
XlsxWriter
bottleneck
numexpr
numba
